+++ src/main/java/org/zendesk/client/v2/model/events/VoiceCommentEvent.java
@@ -16,7 +16,7 @@ public class VoiceCommentEvent extends PublicPrivateEvent {
     private String body;
     private String htmlBody;
     private Boolean trusted;
-    private Integer authorId;
+    private Long authorId;
     private List<String> attachments;
 
     public List<String> getAttachments() {
@@ -28,11 +28,11 @@ public class VoiceCommentEvent extends PublicPrivateEvent {
     }
 
     @JsonProperty("author_id")
-    public Integer getAuthorId() {
+    public Long getAuthorId() {
         return authorId;
     }
 
-    public void setAuthorId(Integer authorId) {
+    public void setAuthorId(Long authorId) {
         this.authorId = authorId;
     }
